# swagger.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/token":{"post":{"summary":"Login","operationId":"login_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"summary":"Read Users Me","operationId":"read_users_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/cards":{"get":{"summary":"Get Cards","operationId":"get_cards_cards_get","parameters":[{"required":true,"schema":{"title":"Offset","type":"integer"},"name":"offset","in":"query"},{"required":true,"schema":{"title":"Cards Amount","type":"integer"},"name":"cards_amount","in":"query"},{"required":true,"schema":{"title":"Money Min","type":"integer"},"name":"money_min","in":"query"},{"required":true,"schema":{"title":"Money Max","type":"integer"},"name":"money_max","in":"query"},{"required":true,"schema":{"title":"Min Capacity","type":"integer"},"name":"min_capacity","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Card"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_token_post":{"title":"Body_login_token_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"Card":{"title":"Card","required":["min_capacity","cost","url","address","tags","name","description"],"type":"object","properties":{"min_capacity":{"title":"Min Capacity","type":"integer"},"cost":{"title":"Cost","type":"integer"},"url":{"title":"Url","type":"string"},"address":{"title":"Address","type":"string"},"tags":{"title":"Tags","type":"array","items":{"type":"string"}},"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}